+++ src/views/Tasks/Tasks.styles.css
@@ -140,7 +140,6 @@
 
   #tasks .filters {
     display: flex !important;
-    flex-direction: column;
     align-items: center;
   }
