+++ src/views/Tasks/components/ListCardsSection/RenderListCards/PaintingCardsList.jsx
@@ -9,7 +9,7 @@ export const PaintingCardsList = ({ renderListTasks, isPhone }) => {
       ) : (
         <>
           {threecolumnListCards.map((columnList) => (
-            <div key={columnList.nameType} className="list">
+            <div key={columnList.txtDiv} className="list">
               <div>{columnList.txtDiv}</div>
               {renderSeparateCards(columnList.nameType)}
             </div>
